finished footer tabs setup and moved my tests to one folder
