+++ test/app.test.js
@@ -1,45 +1,16 @@
-const request = require('supertest');
-const app = require('../server/app.js').app;
-const mongoose = require('../server/app.js').mongoose;
+import React from 'react';
+import { shallow } from 'enzyme';
+import App from '../client/App.jsx';
 
-afterAll(() => {
-  mongoose.connection.close();
-});
+describe('App', () => {
+  test('should render a <div />', () => {
+    const wrapper = shallow(<App />);
+    expect(wrapper).toMatchSnapshot();
+  });
 
-describe('Test the root path', () => {
-  test(
-    'It should response the GET method',
-    async done => {
-      await request(app)
-        .get('/')
-        .then(response => {
-          expect(response.statusCode).toBe(200);
-          done();
-        });
-    },
-    10000
-  );
-});
-
-describe('Test restful api', () => {
-  test(
-    'It should respond with data length equal to 4',
-    async () => {
-      await request(app)
-        .get('/api')
-        .then(res => {
-          expect(res.statusCode).toBe(200);
-        });
-    },
-    10000
-  );
-
-  test('It should respond with an array of objects', async () => {
-    await request(app)
-      .get('/api')
-      .then(res => res.body)
-      .then(data => {
-        expect(Array.isArray(data)).toBe(true);
-      }, 10000);
+  test('should render 4 children elements', () => {
+    const wrapper = shallow(<App />);
+    const wrapperDiv = wrapper.find('div').first();
+    expect(wrapperDiv.children().length).toEqual(4);
   });
 });
